// SmartBalance/Version 1.0.0/Code/Frontend/mersal-accounting-app/src/app/BusinessFeatures/ClosedMonth/Models/closed-month.model.ts
export class ClosedMonth {

    id: any;

    month: any;
    from: any;
    to: any;
    isClosed: boolean;
    closedByUserId: any;
    name: string;
    description: string;

    nameAr: string;
    nameEn: string;

    descriptionAr: string;
    descriptionEn: string;
}

export class ClosedMonthLight {
    
}